Move the FileService interface tests to an test interface
